requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
replicate>=0.34.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
httpx>=0.27.0
hyperframe>=6.0.1
h2>=4.1.0
Pillow>=10.0.0
requests>=2.31.0 